refactor(app): move components to form folder
