+++ src/components/form/review-form.tsx
@@ -6,11 +6,11 @@ import { z } from "zod";
 import dayjs from "dayjs";
 import { Button } from "~/components/ui/button";
 import { Form } from "~/components/ui/form";
-import { ReviewSection } from "~/components/review-section";
-import { CoopCycleSection } from "~/components/coop-cycle-section";
-import { CompanyDetailsSection } from "~/components/company-details-section";
-import { RatingsSection } from "~/components/ratings-section";
-import { SubmissionConfirmation } from "~/components/submission-confirmation";
+import { ReviewSection } from "~/components/form/review-section";
+import { CoopCycleSection } from "~/components/form/coop-cycle-section";
+import { CompanyDetailsSection } from "~/components/form/company-details-section";
+import { RatingsSection } from "~/components/form/ratings-section";
+import { SubmissionConfirmation } from "~/components/form/submission-confirmation";
 import { useState } from "react";
 import { Company, WorkEnvironment, WorkTerm } from "@prisma/client";
 import { api } from "~/trpc/react";
