+++ src/components/form/submission-confirmation.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import Image from "next/image";
 import Link from "next/link";
